requirements.txt
apipkg==1.5
argon2-cffi==20.1.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
confusable-homoglyphs==3.2.0
coverage==5.4
django-bootstrap-form==3.4
django-ckeditor==6.1.0
django-crispy-forms==1.11.0
django-floppyforms==1.9.0
django-js-asset==1.2.2
django-markdownx==3.0.1
django-multiselectfield==0.1.12
django-registration==3.1.1
django-survey-and-report==1.3.33
django-tastypie==0.14.3
djangorestframework==3.12.2
execnet==1.8.0
gdown==4.7.1
iniconfig==1.1.1
Markdown==3.2.1
mysqlclient==2.0.3
openpyxl==3.0.9
ordereddict==1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
PyYAML==5.4.1
reportlab==3.5.59
six==1.15.0
toml==0.10.2